requirements.txt
streamlit==1.39.0
openai==1.54.0
pdfplumber==0.11.4
python-dotenv==1.0.1
tavily-python==0.5.0
httpx==0.27.0
apify-client==1.8.2
requests==2.31.0
beautifulsoup4==4.12.3

